new properties, added warnings and cleaned up syntax

+++ examples/iOS Example/iOS Example/MasterViewController.m
@@ -56,7 +56,7 @@
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
-	return self.menuItems.count;
+	return (NSInteger)self.menuItems.count;
 }
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath
